requirements.txt
Werkzeug==0.11.10
tensorflow==1.7.0
gevent==1.2.2
numpy==1.14.2
Flask==0.11.1
six==1.11.0
Pillow==5.1.0
gunicorn==19.7.1
opencv-python==3.2.0.7
